Refactor tab path display and apply code style cleanup
Replaced magic boolean with configurable integer option for tab path display depth.
Applied consistent code formatting across multiple files, including spacing around control flow statements (if, switch, for) and braces in initializer lists.

Key changes:
* Modified SetTabShowPath() and IsTabShowPath() to use integer count parameter instead of boolean, allowing display of multiple directory levels
* Added GetFileNameWithDirPart() utility function to extract filename with configurable directory depth
* Updated tab label rendering to use new utility function with path depth configuration
* Changed modified file indicator in tabs from asterisk to floppy disk emoji (💾)
* Fixed line-length violations by reformatting long function signatures and calls
* Corrected loop variable types from size_t to int in PatchApplier for consistency with comparison operations

+++ Plugin/Diff/clPatchApplier.cpp
@@ -192,7 +192,7 @@ clStatusOr<int> PatchApplier::ApplyHunk(ITextArea* ctrl, const wxArrayString& li
 
     // First pass: locate the hunk.
     int hunk_found_line{wxNOT_FOUND};
-    for (size_t current_editor_line_number = start_line; current_editor_line_number < totalLines;
+    for (int current_editor_line_number = start_line; current_editor_line_number < totalLines;
          ++current_editor_line_number) {
         wxString editor_line = ctrl->GetLine(current_editor_line_number);
         editor_line.Trim();
@@ -201,7 +201,7 @@ clStatusOr<int> PatchApplier::ApplyHunk(ITextArea* ctrl, const wxArrayString& li
         }
 
         bool hunk_found{true};
-        for (size_t hunk_line_number = 0; hunk_line_number < lines_to_find.size(); ++hunk_line_number) {
+        for (int hunk_line_number = 0; hunk_line_number < lines_to_find.size(); ++hunk_line_number) {
             wxString hunk_line = lines_to_find[hunk_line_number];
             // These lines must exist and match in the control
             if (current_editor_line_number + hunk_line_number >= totalLines) {
